New Update
Cleaned up some stuff including Making sure admin commands can only be run by admins, Moving the delete commands stuff to the CommandBuilderHandler, changed the TTSStateHandlerService to be a static class to make it easily accessible everywhere without dependency injection, fixed the chatgpt integration to defer the response and follow it up after to avoid the slashcommands timeout of 3 seconds from happening when a response is being generated. Implemented a simple check for commands to see if they have been updated or not by name, or options they take, to reduce rebuilding a command every launch which has not changed.

Started adding the foundations and ideas for new commands which have yet to be implemented.
Added kicking command, finished the foundation of the banning command(still need to add some other functionality)
Tested and made sure the chatgpt responses follow the personality I assigned through the system message.