Fixes and Updates
Added the ability for the pickleGPT to dm messages to users with a specific context.
Added a function to send a message to tell every single guild and channel the bot is in, that commands have updated and to restart discord for the newest commands.
Fixed the IsCommandDifferent function IsRequired option so it appropriately checks - Discord only returns True for IsRequired, and if its false it returns a null value, hence this fix compares them appropriately.
Updated MessageDeleted Function for when the cached deleted message cannot be retrieved.